Move netip outside of internal
